+++ test/where_test.rb
@@ -113,31 +113,19 @@ class WhereTest < Minitest::Test
 
   def test_regexp_not_anchored
     store_names ["abcde"]
-    # regular expressions are always anchored right now
-    # TODO change in future release
-    assert_warns "Regular expressions are always anchored in Elasticsearch" do
-      assert_search "*", [], where: {name: /abcd/}
-    end
-    assert_warns "Regular expressions are always anchored in Elasticsearch" do
-      assert_search "*", [], where: {name: /bcde/}
-    end
-    assert_warns "Regular expressions are always anchored in Elasticsearch" do
-      assert_search "*", ["abcde"], where: {name: /abcde/}
-    end
-    assert_warns "Regular expressions are always anchored in Elasticsearch" do
-      assert_search "*", ["abcde"], where: {name: /.*bcd.*/}
-    end
+    assert_search "*", ["abcde"], where: {name: /abcd/}
+    assert_search "*", ["abcde"], where: {name: /bcde/}
+    assert_search "*", ["abcde"], where: {name: /abcde/}
+    assert_search "*", ["abcde"], where: {name: /.*bcd.*/}
   end
 
   def test_regexp_anchored
     store_names ["abcde"]
     assert_search "*", ["abcde"], where: {name: /\Aabcde\z/}
-    assert_warns "Regular expressions are always anchored in Elasticsearch" do
-      assert_search "*", [], where: {name: /\Abcd/}
-    end
-    assert_warns "Regular expressions are always anchored in Elasticsearch" do
-      assert_search "*", [], where: {name: /bcd\z/}
-    end
+    assert_search "*", ["abcde"], where: {name: /\Aabc/}
+    assert_search "*", ["abcde"], where: {name: /cde\z/}
+    assert_search "*", [], where: {name: /\Abcd/}
+    assert_search "*", [], where: {name: /bcd\z/}
   end
 
   def test_regexp_case